Merge PJT6 into master branch

See merge request nts-2020-2nd-team5/jeon_kyung_yoon!31

+++ reservation/src/main/java/kr/or/connect/reservation/controller/api/ProductApiController.java
@@ -55,7 +55,7 @@ public class ProductApiController extends EntityController {
 		@PathVariable(name = "displayInfoId") Long displayInfoId) {
 		DisplayInfo displayInfo = displayService.getDisplayInfoUsingDisplayInfoId(displayInfoId);
 		DisplayInfoImage displayInfoImage = displayService.getDisplayInfoImageUsingDisplayInfoId(displayInfoId);
-		int productId = displayInfo.getProductId();
+		Long productId = displayInfo.getProductId();
 		List<ProductImage> productImages = productService.getProductImagesUsingProductId(productId);
 		List<ProductPrice> productPrices = productService.getProductPricesUsingProductId(productId);
 		List<Comment> comments = commentService.getCommentsUsingProductId(productId);
